Front end design
The code of an Asp.net site. If you want to run and debug the code, you
need to download the Visual Studio 2012 professional (Microsoft offer a
free trial for 90 days):
http://geekswithblogs.net/AbhishekLohani/archive/2013/03/14/visual-studio-test-professional-2012-90-day-trial.aspx.
Just open the entire folder as a website and you will be able to testing
running the site.

I used some server end function from .NET framework because the
predefined module saved a lot of time and efforts other than write code
from scratch. I know you are using PHP which will not work with C#
server end function.